Rename storage package for vsphere impl to vsphere
Renamed the directory in an earlier change without renaming the package.
Fixed for parity;  directory and package now have the same name
(vsphere).
